Conflicts:
	src/Catty.xcodeproj/project.pbxproj
	src/Catty/AppDefines.h
	src/Catty/BrickDetailViewController.m
	src/Catty/FormulaElement.h
	src/Catty/Functions.h
	src/Catty/Functions.m
	src/Catty/Header.m
	src/Catty/LanguageTranslationDefines.h
	src/Catty/Operators.h
	src/Catty/Operators.m

+++ src/Catty/minizip/zip.c
@@ -1374,7 +1374,7 @@ local int zip64FlushWriteBuffer(zip64_internal* zi)
         uInt i;
         int t;
         for (i=0;i<zi->ci.pos_in_buffered_data;i++)
-            zi->ci.buffered_data[i] = zencode(zi->ci.keys, zi->ci.pcrc_32_tab, zi->ci.buffered_data[i],t);
+            zi->ci.buffered_data[i] = (Byte)zencode(zi->ci.keys, zi->ci.pcrc_32_tab, zi->ci.buffered_data[i],t);
 #endif
     }
